requirements.txt
Django>=2.1
#psycopg2>=2.7,<3.0
psycopg2-binary

djangorestframework==3.9.4
markdown==3.1.1
django-filter==2.1.0